Refactor test code for improved readability and consistency
- Reformatted model and config definitions in `test_test_evaluator.py` for better readability.
- Enhanced test function signatures in various test files to improve clarity.
- Updated assertions in `test_training_async_validation.py`, `test_training_class_weights.py`, and others for better formatting.
- Added a new script `find_rich_tables.py` to identify occurrences of specific code patterns.
- Introduced placeholder tests in `test_tuning_dashboard.py` for future implementation of the dashboard module.

+++ src/data/augmentation.py
@@ -401,8 +401,8 @@ def apply_spec_augment_gpu(
             start = int(cp.random.randint(0, mel_bins - mask_width + 1).item())
             spec_gpu[:, start : start + mask_width] = 0
 
-    # Transfer back to CPU
+    # Transfer back to CPU with explicit float32 casting for consistency
     result = cp.asnumpy(spec_gpu)
     del spec_gpu
     cp.get_default_memory_pool().free_all_blocks()
-    return np.asarray(result)
+    return np.asarray(result, dtype=np.float32)
